HZN-1194: Remove unnecessary inheritance
