--- core/jmx/impl/src/main/java/org/opennms/netmgt/jmx/impl/connection/connectors/Jsr160ConnectionWrapper.java
@@ -1,45 +1,0 @@
-/*******************************************************************************
- * This file is part of OpenNMS(R).
- *
- * Copyright (C) 2005-2014 The OpenNMS Group, Inc.
- * OpenNMS(R) is Copyright (C) 1999-2014 The OpenNMS Group, Inc.
- *
- * OpenNMS(R) is a registered trademark of The OpenNMS Group, Inc.
- *
- * OpenNMS(R) is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as published
- * by the Free Software Foundation, either version 3 of the License,
- * or (at your option) any later version.
- *
- * OpenNMS(R) is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with OpenNMS(R).  If not, see:
- *      http://www.gnu.org/licenses/
- *
- * For more information contact:
- *     OpenNMS(R) Licensing <license@opennms.org>
- *     http://www.opennms.org/
- *     http://www.opennms.com/
- *******************************************************************************/
-
-package org.opennms.netmgt.jmx.impl.connection.connectors;
-
-import javax.management.MBeanServerConnection;
-import javax.management.remote.JMXConnector;
-
-/**
- * This class manages the connection to the remote JMX server.  The Jsr160ConnectionFactory
- * class creates the connection and the close method closes it.
- * 
- * @author <A HREF="mailto:mike@opennms.org">Mike Jamison </A>
- */
-public class Jsr160ConnectionWrapper extends DefaultConnectionWrapper {
-    
-    public Jsr160ConnectionWrapper(JMXConnector connector, MBeanServerConnection connection) {
-        super(connector, connection);
-    }
-}
