+++ features/poller/monitors/core/src/main/java/org/opennms/netmgt/poller/monitors/Jsr160Monitor.java
@@ -28,20 +28,7 @@
 
 package org.opennms.netmgt.poller.monitors;
 
-import org.opennms.netmgt.jmx.connection.JmxConnectors;
-
-/**
- * The class is responsible for getting the connection to the rmote jmx server.  The
- * super class (JMXMonitor) performs the checking to see if the service exists and 
- * how long it took to make the connection.
- *  
- * @author <A HREF="mailto:mike@opennms.org">Mike Jamison </A>
- * @author <A HREF="http://www.opennms.org/">OpenNMS </A>
- */
+// It's only purpose is backwards compatibility. Should be removed at some point
 public class Jsr160Monitor extends JMXMonitor {
 
-    @Override
-    protected JmxConnectors getConnectionName() {
-        return JmxConnectors.jsr160;
-    }
 }
